+++ bin/c-yarn.js
@@ -7,8 +7,7 @@ const { missingCommand } = require('./lib/c');
 
 // The basic program, which uses sub-commands.
 program
-    .command('auth', 'Retrieves your NPM auth key from ~/.npmrc')
-    .command('proxy <location> [cmd...]', 'Run an NPM command against an NPM location.')
+    .command('proxy <location> [cmd...]', 'Run a Yarn command against an NPM location.')
     .parse(process.argv);
 
 missingCommand(program);